add map zoom, first pass
git-svn-id: https://svn.code.sf.net/p/triplea/code/trunk/triplea@1533 a4c711aa-641f-0410-879d-9b464e52b682

+++ src/games/strategy/triplea/ui/MapData.java
@@ -367,7 +367,7 @@ public class MapData
     /**
      * Get the territory at the x,y co-ordinates could be null.
      */
-    public String getTerritoryAt(int x, int y)
+    public String getTerritoryAt(double x, double y)
     {
         String seaName = null;
